refactor: made the project follow idiomatic project stucutre
