+++ test/go.mod
@@ -1,0 +1,3 @@
+module test
+
+go 1.21.3
