+++ test/integration_test.go
@@ -1,14 +1,27 @@
-package main
+package test
 
 import (
 	"os"
-	"os/exec"
 	"strings"
 	"testing"
 
 	html "github.com/Agile-Arch-Angels/decentralized-tech-radar_dev/HTML"
 	Reader "github.com/Agile-Arch-Angels/decentralized-tech-radar_dev/SpecReader"
 )
+
+// Tests
+// Integration test
+func TestReaderAndWriter(t *testing.T) {
+	// Set up
+	createCsvFile()
+	defer cleanUp()
+
+	// Read test file
+	specs := Reader.ReadCsvSpec(testFileName + ".csv")
+	html.GenerateHtml(specs)
+
+	assertIndexHTML(t)
+}
 
 // Test Set up
 var testFileName string = "ForTesting"
@@ -79,49 +92,4 @@ func assertIndexHTML(t *testing.T) {
 	if !strings.Contains(contentStr, correctHTML) {
 		t.Errorf("HTML doesn't contain the expected data\nContained:\n%s", contentStr)
 	}
-}
-
-// Tests
-// Integration test
-func TestReaderAndWriter(t *testing.T) {
-	// Set up
-	createCsvFile()
-	defer cleanUp()
-
-	// Read test file
-	specs := Reader.ReadCsvSpec(testFileName + ".csv")
-	html.GenerateHtml(specs)
-
-	assertIndexHTML(t)
-}
-
-// End-to-end test
-func TestEndToEnd(t *testing.T) {
-	// Set up
-	createCsvFile()
-	defer cleanUp()
-
-	// Read test file
-	specs := Reader.ReadCsvSpec(testFileName + ".csv")
-	html.GenerateHtml(specs)
-
-	// Start program using CLI arguments
-	os.Args = []string{"cmd", testFileName + ".csv"}
-	//Works on Unix and Windows
-	cmd := exec.Command("go", "build", "-o", "tech_radar.exe")
-	cmd1 := exec.Command("./tech_radar.exe", "-file", testFileName+".csv")
-
-	_, err := cmd.Output()
-	if err != nil {
-		t.Fatalf("%v", err)
-	}
-
-	cmd1_output, err := cmd1.Output()
-	if err != nil {
-		t.Fatalf("%v", err)
-	} else if !strings.Contains(string(cmd1_output), "Opened csv file!") {
-		t.Errorf("Output didn't match expected. %s", string(cmd1_output))
-	}
-
-	assertIndexHTML(t)
 }
